+++ src/Schema/Directives/GuardDirective.php
@@ -1,0 +1,136 @@
+<?php
+
+namespace Nuwave\Lighthouse\Schema\Directives;
+
+use Closure;
+use GraphQL\Language\AST\TypeDefinitionNode;
+use GraphQL\Language\AST\TypeExtensionNode;
+use Illuminate\Contracts\Auth\Factory as AuthFactory;
+use Nuwave\Lighthouse\Exceptions\AuthenticationException;
+use Nuwave\Lighthouse\Schema\AST\ASTHelper;
+use Nuwave\Lighthouse\Schema\AST\DocumentAST;
+use Nuwave\Lighthouse\Schema\Values\FieldValue;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
+use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
+use Nuwave\Lighthouse\Support\Contracts\TypeExtensionManipulator;
+use Nuwave\Lighthouse\Support\Contracts\TypeManipulator;
+
+/**
+ * @see \Illuminate\Auth\Middleware\Authenticate
+ */
+class GuardDirective extends BaseDirective implements FieldMiddleware, TypeManipulator, TypeExtensionManipulator, DefinedDirective
+{
+    /**
+     * @var \Illuminate\Contracts\Auth\Factory
+     */
+    protected $auth;
+
+    /**
+     * @param  \Illuminate\Contracts\Auth\Factory  $auth
+     */
+    public function __construct(AuthFactory $auth)
+    {
+        $this->auth = $auth;
+    }
+
+    /**
+     * Name of the directive.
+     *
+     * @return string
+     */
+    public function name(): string
+    {
+        return 'guard';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Run authentication through one more guards.
+This is run per field and may allow unauthenticated
+users to still receive partial results.
+"""
+directive @guard(
+  """
+  Specify which guards to use, e.g. "api".
+  When not defined, the default driver is used.
+  """
+  with: [String!]
+) on FIELD_DEFINITION | OBJECT
+SDL;
+    }
+
+    /**
+     * Resolve the field directive.
+     *
+     * @param  \Nuwave\Lighthouse\Schema\Values\FieldValue  $fieldValue
+     * @param  \Closure  $next
+     * @return \Nuwave\Lighthouse\Schema\Values\FieldValue
+     */
+    public function handleField(FieldValue $fieldValue, Closure $next): FieldValue
+    {
+        $resolver = $fieldValue->getResolver();
+
+        return $next(
+            $fieldValue->setResolver(
+                function () use ($resolver) {
+                    $this->authenticate(
+                        (array) $this->directiveArgValue('with')
+                    );
+
+                    return $resolver(...func_get_args());
+                }
+            )
+        );
+    }
+
+    /**
+     * Determine if the user is logged in to any of the given guards.
+     *
+     * @param  string[]  $guards
+     * @return void
+     *
+     * @throws \Illuminate\Auth\AuthenticationException
+     */
+    protected function authenticate(array $guards): void
+    {
+        if (empty($guards)) {
+            $guards = [null];
+        }
+
+        foreach ($guards as $guard) {
+            if ($this->auth->guard($guard)->check()) {
+                $this->auth->shouldUse($guard);
+                return;
+            }
+        }
+
+        $this->unauthenticated($guards);
+    }
+
+    /**
+     * Handle an unauthenticated user.
+     *
+     * @param  array<string|null>  $guards
+     * @return void
+     *
+     */
+    protected function unauthenticated(array $guards): void
+    {
+        throw new AuthenticationException(
+            AuthenticationException::UNAUTHENTICATED,
+            $guards
+        );
+    }
+
+    public function manipulateTypeDefinition(DocumentAST &$documentAST, TypeDefinitionNode &$typeDefinition): void
+    {
+        ASTHelper::addDirectiveToFields($this->directiveDefinition(), $typeDefinition);
+    }
+
+    public function manipulateTypeExtension(DocumentAST &$documentAST, TypeExtensionNode &$typeExtension): void
+    {
+        ASTHelper::addDirectiveToFields($this->directiveDefinition(), $typeExtension);
+    }
+}